Update error and wait functions

+++ src/wait.c
@@ -6,7 +6,7 @@
 /*   By: lkilpela <lkilpela@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 13:41:17 by lkilpela          #+#    #+#             */
-/*   Updated: 2024/03/07 10:07:39 by lkilpela         ###   ########.fr       */
+/*   Updated: 2024/03/07 10:29:21 by lkilpela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,9 +31,9 @@ int	wait_children(t_pipex *p, t_tokenize *t)
 		if (pid == -1)
 			error(ERR_WAITPID);
 		else if (WIFEXITED(status))
-			ft_printf("Child with PID %d exited with status %d.\n", pid, WEXITSTATUS(status));
+			ft_printf("Child with PID %u exited with status %d.\n", (unsigned int)pid, WEXITSTATUS(status));
 		else if (WIFSIGNALED(status))
-			ft_printf("Child with PID %d was terminated by signal %d.\n", pid, WTERMSIG(status));
+			ft_printf("Child with PID %u was terminated by signal %d.\n", (unsigned int)pid, WTERMSIG(status));
 		i++;
 	}
 	return (0);
